requirements.txt
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click==7.1.2
    # via -r requirements.in
idna==2.10
    # via requests
jinja2==2.11.3
    # via -r requirements.in
markupsafe==1.1.1
    # via jinja2
packaging==20.9
    # via -r requirements.in
prettytable==2.1.0
    # via -r requirements.in
pycurl==7.43.0.6
    # via -r requirements.in
pyparsing==2.4.7
    # via packaging
python-magic==0.4.22
    # via -r requirements.in
pyyaml==5.4.1
    # via -r requirements.in
requests==2.25.1
    # via -r requirements.in
urllib3==1.26.4
    # via requests
verboselogs==1.7
    # via -r requirements.in
wcwidth==0.2.5
    # via prettytable
#######################################################################################
# Copyright ETSI Contributors and Others.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#    http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
# implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#######################################################################################
